move reranker module to utils
