+++ KoPrivateGPT/llm/rerank.py
@@ -5,7 +5,7 @@ import openai
 from KoPrivateGPT.DB.base import BaseDB
 from KoPrivateGPT.llm.base import BaseLLM
 from KoPrivateGPT.llm.basic import BasicLLM
-from KoPrivateGPT.reranker.base import BaseReranker
+from KoPrivateGPT.utils.reranker.base import BaseReranker
 from KoPrivateGPT.retrieval.base import BaseRetrieval
 from KoPrivateGPT.schema import Passage
